requirements.txt
django<2
djangorestframework
django-filter
django-cors-headers
